requirements.txt
django~=3.2.0
django-colorfield==0.8.0
django-filter~=22.1.0
djangorestframework~=3.12.0
djoser~=2.1.0
drf-yasg~=1.21.5
Pillow~=9.4.0
